FaceMask_Detection/deploy/requirements.txt
streamlit
numpy
tensorflow-cpu
pillow
keras